+++ tests/test_api.py
@@ -32,6 +32,28 @@ def test_init_production(app_key):
     assert bb.base_url == "https://api.bigbuy.eu/rest"
 
 
+def test_init_sandbox_true(app_key):
+    bb = BigBuy(app_key, sandbox=True)
+    assert bb.base_url == "https://api.sandbox.bigbuy.eu/rest"
+
+    bb = BigBuy(app_key, sandbox=True, mode="production")
+    assert bb.base_url == "https://api.sandbox.bigbuy.eu/rest"
+
+    bb = BigBuy(app_key, sandbox=True, mode="sandbox")
+    assert bb.base_url == "https://api.sandbox.bigbuy.eu/rest"
+
+
+def test_init_sandbox_false(app_key):
+    bb = BigBuy(app_key, sandbox=False)
+    assert bb.base_url == "https://api.bigbuy.eu/rest"
+
+    bb = BigBuy(app_key, sandbox=False, mode="production")
+    assert bb.base_url == "https://api.bigbuy.eu/rest"
+
+    bb = BigBuy(app_key, sandbox=False, mode="sandbox")
+    assert bb.base_url == "https://api.bigbuy.eu/rest"
+
+
 def test_repr_no_app_key():
     bb = BigBuy()
     assert repr(bb) == f"<Bigbuy>"